chore(middleware-recursion-detection): split into modular components
