+++ packages/middleware-recursion-detection/src/configuration.ts
@@ -1,0 +1,3 @@
+export interface PreviouslyResolved {
+  runtime: string;
+}
